# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-covariance",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute the population covariance of two equal-length lists.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits dependencies, exactly as specified:\n\n1. `meanA = floatdivide(sum(a), n)` and `meanB = floatdivide(sum(b), n)`;\n2. fold over each `i`: `da = floatsubtract(a[i], meanA)`,\n   `db = floatsubtract(b[i], meanB)`, `p = floatmultiply(da, db)`,\n   `s = floatadd(s, p)` (starting from `0.0`);\n3. `result = floatdivide(s, n)`.\n\n`a` and `b` must be non-empty and of equal length, otherwise the request is\nrejected with `422` before any dependency call. If a dependency is\nunreachable it reports itself degraded (`503`); if a dependency rejects an\ninput it forwards the `422`.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CovarianceResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid input, or a dependency rejected an input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CovarianceResponse": {
        "type": "object",
        "required": [
          "a",
          "b",
          "result"
        ],
        "properties": {
          "a": {
            "type": "object"
          },
          "b": {
            "type": "object"
          },
          "result": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "EvalRequest": {
        "type": "object",
        "required": [
          "a",
          "b"
        ],
        "properties": {
          "a": {
            "type": "object",
            "description": "The first list of numbers. Must be non-empty and the same length as `b`."
          },
          "b": {
            "type": "object",
            "description": "The second list of numbers. Must be non-empty and the same length as `a`."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      }
    }
  }
}